feat: project list data fetching을 useFetch hook으로 분리
